requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
celery==5.6.2
redis==7.3.0
django-redis==6.0.0
channels==4.3.2
channels-redis==4.3.0
django-celery-beat==2.9.0
django-ratelimit==4.1.0
psycopg[binary]==3.3.3
Pillow==12.1.1
python-dotenv==1.2.2
whitenoise==6.12.0
httpx==0.28.1
beautifulsoup4==4.14.3
gunicorn>=23.0,<24
daphne>=4.1,<5
